// client/src/app/models/VwAccountsDatumModel.ts
export interface VwAccountsDatumModel {
    userName: string;
    pageId: string;
    pagePath: string;
    pageCaption: string;
    pageIsReadOnly: string;
    pageIsPublic: string;
    pageDescription: string;
    rowId: string;
    rowCaption: string;
    columnId: string;
    columnCaption: string;
    linkId: string;
    linkCaption: string;
    linkAurl: string;
}